@@ -4,22 +4,25 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "winbuild": "webpack",
+    "test": "",
+    "webpack": "webpack",
     "start": "webpack-dev-server"
   },
-  "author": "",
+  "author": "Christopher Bloom",
   "license": "ISC",
   "devDependencies": {
-    "css-loader": "^0.15.6",
-    "node-sass": "^3.2.0",
-    "sass-loader": "^2.0.0",
-    "style-loader": "^0.12.3",
-    "webpack": "^1.11.0",
-    "webpack-dev-server": "^1.10.1"
+    "@babel/core": "^7.0.0-beta.31",
+    "@babel/preset-env": "^7.0.0-beta.31",
+    "babel-loader": "^8.0.0-beta.0",
+    "css-loader": "^0.28.7",
+    "node-sass": "^4.6.0",
+    "sass-loader": "^6.0.6",
+    "style-loader": "^0.19.0",
+    "webpack": "^3.8.1",
+    "webpack-dev-server": "^2.9.4"
   },
   "dependencies": {
-    "jquery": "^2.1.4",
-    "lodash": "^3.10.1"
+    "jquery": "^3.2.1",
+    "lodash": "^4.17.4"
   }
 }
