@@ -1,7 +1,9 @@
+const path = require('path');
+
 module.exports = {
-  entry: './index.js',
+  entry: path.resolve(__dirname, 'index.js'),
   output: {
     filename: 'bundle.js',
-    path: __dirname
+    path: path.resolve(__dirname),
   }
 };
