@@ -1,7 +1,23 @@
+const path = require('path');
+
 module.exports = {
-  entry: './index.js',
+  entry: path.resolve(__dirname, 'index.js'),
   output: {
     filename: 'bundle.js',
-    path: __dirname
+    path: path.resolve(__dirname),
+  },
+  module: {
+    rules: [
+      {
+        test: /\.js$/,
+        exclude: /(node_modules|bower_components)/,
+        use: {
+          loader: 'babel-loader',
+          options: {
+            presets: ['@babel/preset-env']
+          }
+        }
+      }
+    ]
   }
 };
